refactor: move types file to types folder
