+++ src/types/types.ts
@@ -1,1 +1,3 @@
 export type EntityType = 'user' | 'artist' | 'track' | 'album' | 'favorites';
+
+export type LoggingLevel = 'log' | 'error' | 'warn' | 'debug';
